refactor: move viewport methods to file
